move router to xyz/documentation
